1.21.70 + Major network rework

+++ src/MiNET/MiNET/Particles/DestroyBlockParticle.cs
@@ -25,6 +25,7 @@
 
 using System.Numerics;
 using MiNET.Net;
+using MiNET.Net.Packets.Mcpe;
 using MiNET.Worlds;
 
 namespace MiNET.Particles
@@ -40,7 +41,7 @@ namespace MiNET.Particles
 		public override void Spawn()
 		{
 			McpeLevelEvent particleEvent = McpeLevelEvent.CreateObject();
-			particleEvent.eventId = (int)LevelEventType.ParticlesDestroyBlock;
+			particleEvent.eventId = LevelEventType.ParticlesDestroyBlock;
 			particleEvent.position = Position;
 			particleEvent.data = Data;
 			Level.RelayBroadcast(particleEvent);
